Test refactoring, fixed CHAR
