+++ server/fileno_check_linux.go
@@ -12,10 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build !linux && !darwin
-// +build !linux,!darwin
+package server
 
-package main
+import (
+	"golang.org/x/sys/unix"
+)
 
 func warnIfMaxFilesTooLow() {
+	var lim unix.Rlimit
+	if err := unix.Getrlimit(unix.RLIMIT_NOFILE, &lim); err != nil {
+		return
+	}
+	lim.Cur = lim.Max
+	if err := unix.Setrlimit(unix.RLIMIT_NOFILE, &lim); err != nil {
+		return
+	}
+	if err := unix.Getrlimit(unix.RLIMIT_NOFILE, &lim); err != nil {
+		return
+	}
 }
